moved lenovo fan table
