+++ LenovoLegionToolkit.Lib/Controllers/GodMode/GodModeControllerV2.cs
@@ -297,7 +297,7 @@ public class GodModeControllerV2 : AbstractGodModeController
         if (Log.Instance.IsTraceEnabled)
             Log.Instance.Trace($"Reading fan table data...");
 
-        var data = await WMI.LenovoFanMethod.FanGetTableAsync().ConfigureAwait(false);
+        var data = await WMI.LenovoFanTable.ReadAsync().ConfigureAwait(false);
 
         var fanTableData = data
             .Where(d => d.mode == (int)powerModeState + 1)
