+++ LenovoLegionToolkit.Lib/System/WMI.LenovoFanMethod.cs
@@ -1,8 +1,6 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using LenovoLegionToolkit.Lib.Extensions;
 
 // ReSharper disable InconsistentNaming
 // ReSharper disable StringLiteralTypo
@@ -13,24 +11,10 @@ public static partial class WMI
 {
     public static class LenovoFanMethod
     {
-        public static Task<bool> ExistsAsync(int sensorId, int fanId) => WMI.ExistsAsync("root\\WMI", $"SELECT * FROM LENOVO_FAN_TABLE_DATA WHERE Sensor_ID = {sensorId} AND Fan_Id = {fanId}");
-
-        public static Task<IEnumerable<(int mode, byte fanId, byte sensorId, ushort[] fanTableData, ushort[] sensorTableData)>> FanGetTableAsync() => ReadAsync("root\\WMI",
-            $"SELECT * FROM LENOVO_FAN_TABLE_DATA",
-            pdc =>
-            {
-                var mode = pdc.Contains("Mode") ? Convert.ToInt32(pdc["Mode"].Value) : 0;
-                var fanId = Convert.ToByte(pdc["Fan_Id"].Value);
-                var sensorId = Convert.ToByte(pdc["Sensor_ID"].Value);
-                var fanTableData = (ushort[]?)pdc["FanTable_Data"].Value ?? Array.Empty<ushort>();
-                var sensorTableData = (ushort[]?)pdc["SensorTable_Data"].Value ?? Array.Empty<ushort>();
-                return (mode, fanId, sensorId, fanTableData, sensorTableData);
-            });
-
         public static Task FanSetTableAsync(byte[] fanTable) => CallAsync("root\\WMI",
-            $"SELECT * FROM LENOVO_FAN_METHOD",
-            "Fan_Set_Table",
-            new() { { "FanTable", fanTable } });
+        $"SELECT * FROM LENOVO_FAN_METHOD",
+        "Fan_Set_Table",
+        new() { { "FanTable", fanTable } });
 
         public static Task<bool> FanGetFullSpeedAsync() => CallAsync("root\\WMI",
             $"SELECT * FROM LENOVO_FAN_METHOD",
